Code factor & small issues.
