+++ WalletWasabi/Tor/Control/Messages/CircuitStatus/CircPath.cs
@@ -1,0 +1,12 @@
+namespace WalletWasabi.Tor.Control.Messages.CircuitStatus
+{
+	/// <summary>	
+	/// </summary>
+	/// <remarks>
+	/// <see cref="FingerPrint"/> is a 40*HEXDIG string.
+	/// <para><see cref="Nickname"/> matches <c>^[a-zA-Z0-9]{1, 19}$</c>.</para>
+	/// </remarks>
+	public record CircPath(string FingerPrint, string? Nickname)
+	{
+	}
+}
